feat: 修改停止状态下的视频播放、跳转实现

+++ Src/MCVideoPlayTest/MCVideoPlayTest.h
@@ -21,13 +21,13 @@ public:
 private slots:
     void timerTimeOut();
     void videoPlayOrStop();
-    void videoSeek(qint64 seekTime);
-    void videoDurationChanged(qint64 msecond);
+    void videoSeek(int seekTime);
+    void videoDurationChanged(int msecond);
     void videoFrameChanged(std::shared_ptr<MCWidget::MCVideoFrame> frame);
     void videoStateChanged(MCWidget::MCVideoPlayer::VideoState state);
 
 private:
-    QString getTimeString(qint64 msecond);
+    QString getTimeString(int msecond);
 
 private:
     Ui::MCVideoPlayTest*        ui;
